docs: refactor all organization of documentation
chore: update readme

chore: format and clean
