+++ docs/cli/.pages
@@ -1,0 +1,5 @@
+nav:
+  - AutoPipeline.md
+  - nnUNet.md
+  - imgtools.md
+  - ...
