+++ docs/cli/AutoPipeline.md
