+++ src/imgtools/dicom/index/models/__init__.py
@@ -1,0 +1,9 @@
+from .models import Image, Patient, Series, Study, mapper_registry
+
+__all__ = [
+	'Image',
+	'Patient',
+	'Series',
+	'Study',
+	'mapper_registry',
+]
