+++ src/imgtools/logging/logging_config.py
@@ -22,7 +22,7 @@ class LoggingManager:
 	Manages the configuration and initialization of a structured logger.
 
 	This class provides flexible options for configuring log levels, formats,
-	and output destinations. 
+	and output destinations.
 
 	Parameters
 	----------
@@ -75,13 +75,10 @@ class LoggingManager:
 
 	@property
 	def env_level(self) -> str:
-		return os.environ.get(
-			f'{self.name}_LOG_LEVEL'.upper(), 
-			DEFAULT_LOG_LEVEL
-		).upper()
+		return os.environ.get(f'{self.name}_LOG_LEVEL'.upper(), DEFAULT_LOG_LEVEL).upper()
 
 	@property
-	def base_logging_config(self ) -> Dict:
+	def base_logging_config(self) -> Dict:
 		"""
 		Create the basic logging configuration settings.
 
@@ -97,7 +94,7 @@ class LoggingManager:
 				'console': {
 					'()': structlog.stdlib.ProcessorFormatter,
 					'processors': [
-						CallPrettifier(concise=True), 
+						CallPrettifier(concise=True),
 						structlog.stdlib.ProcessorFormatter.remove_processors_meta,
 						structlog.dev.ConsoleRenderer(
 							exception_formatter=structlog.dev.RichTracebackFormatter(
@@ -146,7 +143,7 @@ class LoggingManager:
 		"""
 		Initialize the logger with the current configuration.
 		"""
-	
+
 		logging.config.dictConfig(self.base_logging_config)
 		structlog.configure(
 			processors=[
@@ -169,45 +166,39 @@ class LoggingManager:
 		"""
 		return structlog.get_logger(self.name)
 
-	def configure_logging(
-			self, level: str = DEFAULT_LOG_LEVEL
-	) -> structlog.stdlib.BoundLogger:
-			"""
-			Dynamically adjust logging settings.
+	def configure_logging(self, level: str = DEFAULT_LOG_LEVEL) -> structlog.stdlib.BoundLogger:
+		"""
+		Dynamically adjust logging settings.
 
-			Parameters
-			----------
-			level : str, optional
-					Set the log level.
+		Parameters
+		----------
+		level : str, optional
+				Set the log level.
 
-			Returns
-			-------
-			structlog.stdlib.BoundLogger
-					Updated logger instance.
+		Returns
+		-------
+		structlog.stdlib.BoundLogger
+				Updated logger instance.
 
-			Raises
-			------
-			ValueError
-					If an invalid log level is specified.
-			"""
-			level_upper = level.upper()
-			if level_upper not in VALID_LOG_LEVELS:
-					msg = f'Invalid logging level: {level}'
-					raise ValueError(msg)
-			
-			# Store the old level for logging the change
-			old_level = self.level
-			self.level = level_upper
+		Raises
+		------
+		ValueError
+				If an invalid log level is specified.
+		"""
+		level_upper = level.upper()
+		if level_upper not in VALID_LOG_LEVELS:
+			msg = f'Invalid logging level: {level}'
+			raise ValueError(msg)
 
-			self._initialize_logger()
-			logger = self.get_logger()
-			
-			# Log the level change
-			if old_level != self.level:
-					logger.info(
-							"Log level changed",
-							old_level=old_level,
-							new_level=self.level
-					)
+		# Store the old level for logging the change
+		old_level = self.level
+		self.level = level_upper
 
-			return logger
+		self._initialize_logger()
+		logger = self.get_logger()
+
+		# Log the level change
+		if old_level != self.level:
+			logger.info('Log level changed', old_level=old_level, new_level=self.level)
+
+		return logger
